Extract shared filter logic into useFilterState composable
The FilterSidebar and MobileFilterSheet components had a lot of duplicated
filter state management code. This pulls that out into a shared composable,
cutting down on repetition and making both components simpler to follow.
